Refactor Settings into a folder; fix layout, transition, go asset-only
Split the two large Settings files into Sources/md/Settings/ (AppSettings,
SettingsWindowController, GeneralSettingsView, AppearanceSettingsView,
FontSettings, SettingsControls, AntialiasingText), each with a header comment.

UI fixes:
- Restore CotEditor-style top padding; keep extra side/bottom breathing room.
- Stop auto-focusing (and ringing) the first control via .focusEffectDisabled().
- Pin the font-display field to a fixed height (it collapsed when derived from a
  not-yet-laid-out frame).
- Smooth the pane-switch transition: swap content, measure the top-pinned target
  frame, and animate only the window frame (no fade, no competing resize).

Drop the in-app accent swatches and highlight-color picker in favor of the
AccentColor asset (matches CotEditor): native controls follow the asset / system
accent, so the custom checkbox/radio styles and tint plumbing are gone.
