+++ Sources/md/Settings/AntialiasingText.swift
@@ -22,10 +22,10 @@ struct AntialiasingText: NSViewRepresentable {
         nsView.allowsExpansionToolTips = true
         nsView.setContentCompressionResistancePriority(.defaultLow, for: .horizontal)
 
-        // Keep the field at a stable height (a couple points taller than the
-        // default so a 16pt preview has a little breathing room) regardless of
-        // the previewed font's size.
-        nsView.heightAnchor.constraint(equalToConstant: nsView.frame.height + 4).isActive = true
+        // Pin a fixed, stable height so a 16pt preview fits with a little
+        // breathing room. (Deriving it from `frame.height` collapses the field —
+        // the frame is zero-height before Auto Layout has sized it.)
+        nsView.heightAnchor.constraint(equalToConstant: 24).isActive = true
 
         return nsView
     }
